// src/app/models/student.ts
export class Student{
    studId: number;
    studFirstName: string;
    studMiddleName: string;
    studLastName: string;
    studAddressLine: string;
    studRegion: string;
    studCity: string;
    studBrgy: string;
    studZipCode: string;
    studNationality: string;
    studAge: string;
    studGender: string;
    studMaritalStatus: string;
    studDob: Date;
    studDln: string;
    studPhoto: string;
    studSuffix: string;
    studProvince: string;
    studContactNo: string;
    studStatus: string;
    studLtoPortalId: string;
    studEmail: String;
}